requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
chardet==3.0.4
decorator==4.0.11
discord.py==0.16.8
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
multidict==3.0.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
PyYAML==3.12
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
websockets==3.3
